pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abhijit Gupta <avgupta456@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.14.0"
Flask = "^1.1.2"
openai = "^0.2.4"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
